+++ src/Ether.Network/INetClient.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using Ether.Network.Packets;
+
+namespace Ether.Network
+{
+    public interface INetClient
+    {
+        Guid Id { get; }
+        bool IsConnected { get; }
+        void Connect();
+        void Disconnect();
+        void Send(NetPacketBase packet);
+    }
+}